Separerer ut det å gå frå mal til pdf i biblioteksmetode
