+++ brevbaker/src/main/kotlin/no/nav/brev/brevbaker/BrevbakerJSON.kt
@@ -1,0 +1,9 @@
+package no.nav.brev.brevbaker
+
+import no.nav.pensjon.brev.api.model.maler.BrevbakerBrevdata
+import no.nav.pensjon.brev.template.Letter
+import no.nav.pensjon.brev.template.render.Letter2Markup
+
+object BrevbakerJSON {
+    fun renderJSON(letter: Letter<BrevbakerBrevdata>) = Letter2Markup.render(letter).letterMarkup
+}
